cli refactor: move dvc.command.base to dvc.cli.command
