+++ dvc/command/check_ignore.py
@@ -1,7 +1,8 @@
 import argparse
 
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link
 from dvc.command import completion
-from dvc.command.base import CmdBase, append_doc_link
 from dvc.ui import ui
