+++ dvc/command/diff.py
@@ -2,8 +2,9 @@ import argparse
 import logging
 import os
 
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link
 from dvc.command import completion
-from dvc.command.base import CmdBase, append_doc_link
 from dvc.ui import ui
 
 logger = logging.getLogger(__name__)
