+++ dvc/command/experiments/init.py
@@ -3,7 +3,8 @@ import logging
 
 from funcy import compact
 
-from dvc.command.base import CmdBase, append_doc_link
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link
 from dvc.exceptions import InvalidArgumentError
 from dvc.ui import ui
