+++ dvc/command/get_url.py
@@ -1,10 +1,11 @@
 import argparse
 import logging
 
+from dvc.cli.command import CmdBaseNoRepo
 from dvc.exceptions import DvcException
 
+from ..cli.utils import append_doc_link
 from . import completion
-from .base import CmdBaseNoRepo, append_doc_link
 
 logger = logging.getLogger(__name__)
