+++ dvc/command/params.py
@@ -1,8 +1,9 @@
 import argparse
 import logging
 
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link, fix_subparsers
 from dvc.command import completion
-from dvc.command.base import CmdBase, append_doc_link, fix_subparsers
 from dvc.exceptions import DvcException
 from dvc.ui import ui
