+++ dvc/command/plots.py
@@ -4,8 +4,9 @@ import logging
 
 from funcy import first
 
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link, fix_subparsers
 from dvc.command import completion
-from dvc.command.base import CmdBase, append_doc_link, fix_subparsers
 from dvc.exceptions import DvcException
 from dvc.ui import ui
 from dvc.utils import format_link
