+++ dvc/command/run.py
@@ -1,7 +1,8 @@
 import argparse
 import logging
 
-from dvc.command.base import CmdBase, append_doc_link
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link
 from dvc.command.stage import parse_cmd
 from dvc.exceptions import DvcException
