+++ dvc/command/stage.py
@@ -3,8 +3,9 @@ import logging
 from itertools import chain, filterfalse
 from typing import TYPE_CHECKING, Dict, Iterable, List
 
+from dvc.cli.command import CmdBase
+from dvc.cli.utils import append_doc_link, fix_subparsers
 from dvc.command import completion
-from dvc.command.base import CmdBase, append_doc_link, fix_subparsers
 from dvc.utils.cli_parse import parse_params
 from dvc.utils.humanize import truncate_text
